Remove deprecated components and update package-lock.json to reflect changes in dependency management.
