+++ app/(website)/about/about.tsx
@@ -1,0 +1,7 @@
+export default function AboutPage() {
+    return (
+        <div>
+            <h1>About</h1>
+        </div>
+    )
+}
