+++ app/dashboard/page.tsx
@@ -1,0 +1,7 @@
+export default function DashboardPage() {
+    return (
+        <div>
+            <h1>Dashboard</h1>
+        </div>
+    )
+}
