+++ components/website/how-it-works.tsx
